batch: Move discard application
merge.py handled forward batch merge and discard reversal. The discard side
pulled in baseline correspondence, realized boundary lookup, and reversal
helpers that are not part of forward merge placement.

The new batch.discard module owns discard buffer construction, provenance
entries, and absence restoration. Text plan builders now import discard and
merge paths from separate modules.

This commit moves the discard application path out of merge.py.

Forward merge code now keeps candidate enumeration and presence or absence
enforcement without the discard-only dependency chain.
